+++ ldap2pg/__init__.py
@@ -1,6 +1,10 @@
 from .config import __version__, __dist__
+from .utils import UserError
+from .script import synchronize
 
 __all__ = [
+    'UserError',
     '__dist__',
     '__version__',
+    'synchronize',
 ]