+++ tests/unit/test_script.py
@@ -4,9 +4,9 @@ import pytest
 
 
 def test_main(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
-    wm = mocker.patch('ldap2pg.script.wrapped_main', autospec=True)
-    wm.return_value = 0
+    mocker.patch('ldap2pg.script.Configuration')
+    s = mocker.patch('ldap2pg.script.synchronize', autospec=True)
+    s.return_value = 0
 
     from ldap2pg.script import main
 
@@ -17,12 +17,12 @@ def test_main(mocker):
 
 
 def test_bdb_quit(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
-    w = mocker.patch('ldap2pg.script.wrapped_main')
+    mocker.patch('ldap2pg.script.Configuration')
+    s = mocker.patch('ldap2pg.script.synchronize')
 
     from ldap2pg.script import main, pdb
 
-    w.side_effect = pdb.bdb.BdbQuit()
+    s.side_effect = pdb.bdb.BdbQuit()
 
     with pytest.raises(SystemExit) as ei:
         main()
@@ -31,12 +31,12 @@ def test_bdb_quit(mocker):
 
 
 def test_unhandled_error(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
-    w = mocker.patch('ldap2pg.script.wrapped_main')
+    mocker.patch('ldap2pg.script.Configuration')
+    s = mocker.patch('ldap2pg.script.synchronize')
 
     from ldap2pg.script import main
 
-    w.side_effect = Exception()
+    s.side_effect = Exception()
 
     with pytest.raises(SystemExit) as ei:
         main()
@@ -45,12 +45,12 @@ def test_unhandled_error(mocker):
 
 
 def test_user_error(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
-    w = mocker.patch('ldap2pg.script.wrapped_main')
+    mocker.patch('ldap2pg.script.Configuration')
+    s = mocker.patch('ldap2pg.script.synchronize')
 
     from ldap2pg.script import main, UserError
 
-    w.side_effect = UserError("Test message.", exit_code=0xCAFE)
+    s.side_effect = UserError("Test message.", exit_code=0xCAFE)
 
     with pytest.raises(SystemExit) as ei:
         main()
@@ -59,12 +59,12 @@ def test_user_error(mocker):
 
 
 def test_pdb(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
+    mocker.patch('ldap2pg.script.Configuration')
     mocker.patch('ldap2pg.script.os.environ', {'DEBUG': '1'})
     isatty = mocker.patch('ldap2pg.script.sys.stdout.isatty')
     isatty.return_value = True
-    w = mocker.patch('ldap2pg.script.wrapped_main')
-    w.side_effect = Exception()
+    s = mocker.patch('ldap2pg.script.synchronize')
+    s.side_effect = Exception()
     pm = mocker.patch('ldap2pg.script.pdb.post_mortem')
 
     from ldap2pg.script import main
@@ -76,9 +76,8 @@ def test_pdb(mocker):
     assert os.EX_SOFTWARE == ei.value.code
 
 
-def test_wrapped_main(mocker):
+def test_synchronize(mocker):
     from ldap2pg.utils import Timer
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
     PSQL = mocker.patch('ldap2pg.script.PSQL', autospec=True)
     PSQL.return_value.timer = Timer()
     clc = mocker.patch('ldap2pg.script.ldap.connect')
@@ -86,17 +85,17 @@ def test_wrapped_main(mocker):
     manager = SM.return_value
     manager.sync.return_value = 0
 
-    from ldap2pg.script import wrapped_main
+    from ldap2pg.script import synchronize, Configuration
 
-    config = mocker.MagicMock(name='config')
+    config = mocker.MagicMock(name='config', spec=Configuration)
     # Dry run
     config.get.return_value = True
-    wrapped_main(config=config)
+    synchronize(config=config)
 
     # Real mode
     config.get.return_value = False
     PSQL.return_value.return_value = mocker.MagicMock(name='psql')
-    wrapped_main(config=config)
+    synchronize(config=config)
 
     assert clc.called is True
     assert manager.sync.called is True
@@ -104,26 +103,25 @@ def test_wrapped_main(mocker):
     # No LDAP
     clc.reset_mock()
     config.has_ldap_query.return_value = []
-    wrapped_main(config=config)
+    synchronize(config=config)
 
     assert clc.called is False
 
 
-def test_conn_errors(mocker):
-    mocker.patch('ldap2pg.script.dictConfig', autospec=True)
-    mocker.patch('ldap2pg.script.Configuration', autospec=True)
+def test_synchronize_conn_errors(mocker):
+    mocker.patch('ldap2pg.script.Configuration', new=mocker.MagicMock)
     mocker.patch('ldap2pg.script.SyncManager', autospec=True)
     clc = mocker.patch('ldap2pg.script.ldap.connect')
     PSQL = mocker.patch('ldap2pg.script.PSQL', autospec=True)
 
     from ldap2pg.script import (
-        wrapped_main, ConfigurationError,
+        synchronize, ConfigurationError,
         ldap, psycopg2,
     )
 
     clc.side_effect = ldap.LDAPError("pouet")
     with pytest.raises(ConfigurationError):
-        wrapped_main()
+        synchronize()
     clc.side_effect = None
 
     psql = PSQL.return_value
@@ -131,4 +129,12 @@ def test_conn_errors(mocker):
     psql_ = psql.return_value.__enter__.return_value
     psql_.side_effect = psycopg2.OperationalError()
     with pytest.raises(ConfigurationError):
-        wrapped_main()
+        synchronize()
+
+
+def test_init_config_str():
+    from ldap2pg.script import init_config
+
+    config = init_config("""- role: myrole""", environ=dict(), argv=[])
+    assert 1 == len(config['sync_map'])
+    assert 'myrole' in str(config['sync_map'][0]['roles'][0].names[0])