+++ docs/changelog.md
@@ -10,6 +10,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 # Unreleased
 
 - Search for `/etc/openldap/ldap.conf` too.
+- Protect role name when reassigning objects.
 
 
 # ldap2pg 5.5