+++ docs/changelog.md
@@ -11,7 +11,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 - Search for `/etc/openldap/ldap.conf` too.
 - Protect role name when reassigning objects.
-- Apply `LDAPCONF` from environment.
+- Apply `LDAPCONF` `LDAPRC` from environment.
 
 
 # ldap2pg 5.5