+++ docs/changelog.md
@@ -7,7 +7,9 @@ follow [merged Pull request
 pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Amerged).
 
 
-# Unreleased
+# ldap2pg 5.6
+
+Released 2021-06-21.
 
 - Accepts `.lower()` and `.upper()` filters.
 - Rename from lower/upper case to mixed case.